requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.6.0
Flask-CORS==4.0.0
Flask-Bcrypt==1.0.1
marshmallow==3.21.1
marshmallow-sqlalchemy==1.0.0
PyMySQL==1.1.0
python-dotenv==1.0.1
celery==5.3.6
redis==5.0.1
Pillow==10.2.0
boto3==1.34.34
twilio==8.12.0
gunicorn==21.2.0
pytest==8.0.0
pytest-flask==1.3.0
alembic==1.13.1
cryptography==42.0.2
phonenumbers==8.13.30
